Reorganize repo

+++ utils/History.py
@@ -5,8 +5,8 @@ class History(object):
     Accumulator keeping track of the N previous frames to be used by the agent
     for evaluation
     """
-    def __init__(self, shape):
-        self._buffer = np.zeros(shape, dtype=np.float32)
+    def __init__(self, shape, states_dtype=np.float32):
+        self._buffer = np.zeros(shape, dtype=states_dtype)
 
     @property
     def value(self):
